// CelularesNaac/src/app/CelularesNaac/models/dispositivo-movil.ts
export class DispositivoMovil {
    codigo: string;
    cantidad: number;
    marca: string;
    modelo: string;
    precioCompra: number;
    precioVenta: number;
    porcentajeIva: number;
    porcentajeDescuento: number;
    tipoPantalla: string;
    resolucion: string;
    procesador: string;
    almacenamiento: string;
    ram: string;
    camara: string;
    tipoBateria: string;
    capacidadBateria: string;
    tipoProteccion: string;
    lectorHuella: string;
    imagen: string;
}